backend/requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-jazzmin==3.0.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
setuptools>=65.0.0
gunicorn==21.2.0
whitenoise==6.6.0
